user can complete task

+++ src/routes.js
@@ -42,7 +42,45 @@ export const routes = [
     handle: (req, res ) => {
       const { id } = req.params
 
+      if(!id) {
+        return res.writeHead(400).end(JSON.stringify({message: 'please provide a id'}))
+      }
+
       if(database.update('tasks', id, req.body)) {
+        return res.writeHead(204).end()
+      } else {
+        return res.writeHead(400).end(JSON.stringify({message: 'task does not exist'}))
+      }     
+    }
+  },
+  {
+    method: 'PUT',
+    path: buildRoutePath('/tasks/:id'),
+    handle: (req, res ) => {
+      const { id } = req.params
+
+      if(!id) {
+        return res.writeHead(400).end(JSON.stringify({message: 'please provide a id'}))
+      }
+
+      if(database.update('tasks', id, req.body)) {
+        return res.writeHead(204).end()
+      } else {
+        return res.writeHead(400).end(JSON.stringify({message: 'task does not exist'}))
+      }     
+    }
+  },
+  {
+    method: 'PATCH',
+    path: buildRoutePath('/tasks/:id/complete'),
+    handle: (req, res ) => {
+      const { id } = req.params
+
+      if(!id) {
+        return res.writeHead(400).end(JSON.stringify({message: 'please provide a id'}))
+      }
+
+      if(database.complete('tasks', id)) {
         return res.writeHead(204).end()
       } else {
         return res.writeHead(400).end(JSON.stringify({message: 'task does not exist'}))
@@ -54,6 +92,10 @@ export const routes = [
     path: buildRoutePath('/tasks/:id'),
     handle: (req, res ) => {
       const { id } = req.params
+
+      if(!id) {
+        return res.writeHead(400).end(JSON.stringify({message: 'please provide a id'}))
+      }
 
       if(database.delete('tasks', id)) {
         return res.writeHead(204).end()
